+++ lib/annotate/annotate_routes.rb
@@ -4,7 +4,7 @@
 #
 #
 #
-# Prepends the output of "rake routes" to the top of your routes.rb file.
+# Prepends the output of "rails routes" to the top of your routes.rb file.
 # Yes, it's simple but I'm thick and often need a reminder of what my routes
 # mean.
 #